feat(ui): implement ChatInterface component and utility functions
- Introduced a new ChatInterface component to encapsulate chat functionality, improving modularity and readability.
- Added preventDefault utility function to streamline event handling in forms.
- Refactored the main page to utilize the new ChatInterface, enhancing the user experience during conversations.
- Updated API interaction logic to support conversation management and error handling more effectively.
