+++ apps/ui/src/lib/utils.ts
@@ -1,0 +1,6 @@
+export function preventDefault(fn: (event: Event) => void) {
+	return (event: Event) => {
+		event.preventDefault()
+		fn(event)
+	}
+}
